+++ PEedit/x64.h
@@ -1,5 +1,5 @@
 #pragma once	
-extern "C" void* insert_dll(ULONGLONG& size);
-extern "C" void* enter_code(ULONGLONG& size);
-extern "C" void* copy_code(ULONGLONG& size);
-extern "C" void* decode_code(ULONGLONG & size);
+extern "C" void* insert_dll(size_t & size);
+extern "C" void* enter_code(size_t & size);
+extern "C" void* copy_code(size_t & size);
+extern "C" void* decode_code(size_t & size);
